[mp4tree] Break out main
  * Break out program main functionality -> main.c
  * Break out more commmon functionality -> common.*
