+++ mp4tree.h
@@ -4,15 +4,17 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-typedef void (*mp4tree_box_func)
-(
-    const uint8_t * p,
-    size_t          len,
-    int             depth
-);
 
+/* Pointer to a buffer parsing function */
+typedef void (*mp4tree_parse_func) (const uint8_t * p, size_t len, int depth);
+
+/* Print data buffer to stdout hexdump style */
 void
-mp4tree_hexdump(
-    const uint8_t * p,
-    size_t          len,
-    int             depth);
+mp4tree_hexdump(const uint8_t * p, size_t len, int depth);
+
+/* Print MP4 box */
+void
+mp4tree_print(const uint8_t * p, size_t len, int depth);
+
+/* Run self-test */
+int mp4tree_selftest();
